feat(serverHandler): use relative path for links
Prevent using absolute path can make links still work if the server is
behind a reverse http proxy, and the proxy uses a sub path for proxying.

However, when finish uploading files, there is a http redirect send to
client. As redirect location must be an absolute path, It's proxy
owner's responsibility to replace the `location` header with correct
one.