feat: remove option `--bind`
The purpose of providing `--bind` is to provide a possibility to align
the behavior with Windows that path is case-insensitive.

Now just make `--alias` case-insensitive on Windows platform, and remove
the `--bind` option.

The purpose of this change is path related permission controls, like
upload, mkdir, delete and archive, should all align with the case role
based on current platform. This change makes a preparation for that.

BREAKING CHANGE

Before:

Use `--alias` for case-sensitive alias for all platforms.
Use `--bind` for case-insensitive alias for all platforms.

After:

On Windows platform, `--alias` is case-insensitive.
On non-Windows platform, `--alias` is case-sensitive.