feat: replace `--force-dir-slash` with `--auto-dir-slash`
If a file is requested with tailing "/" in the URL,
redirect to the URL without the suffix.

BREAKING CHANGE

Before:

Option name is `--force-dir-slash`.

If a directory list page is requested without tailing "/" in the URL,
redirect to the URL with the suffix.

After:

Option name is `--auto-dir-slash`

If a directory list page is requested without tailing "/" in the URL,
redirect to the URL with the suffix.

If a file is requested with tailing "/" in the URL,
redirect to the URL without the suffix.